Okay maybe it's not working

+++ common/src/main/java/com/ewan/meworking/codec/ClientDataEncoder.java
@@ -32,7 +32,7 @@ public class ClientDataEncoder
         output.flush();
 
         byte[] outArray = outStream.toByteArray();
-        out.writeShort(outArray.length);
+        out.writeInt(outArray.length);
         out.writeBytes(outArray);
     }
 }
